Update README.md
Electricity Billing System is a software-based application developed in Java programming language. The main objective is to manage the details of Electricity Bill, Connections, Store Record of Customers. The project aims at serving the department of electricity by computerizing the billing system. It mainly focuses on the calculation of Units consumed during the specified time and the money to be paid to electricity offices. This computerized system will make the overall billing system easy, accessible, comfortable and effective for consumers.